facenet-prototype/requirements.txt
streamlit
torch
facenet-pytorch
Pillow
